+++ src/components/transaction-inspector/SignTransaction.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
-import SignTransactionButton from '../../components/signTransactionButton';
-import { TxValidationState, VoteValidationState } from '../../components/types/types';
+import SignTransactionButton from './SignTransaction';
+import { TxValidationState, VoteValidationState } from '../../types/types';
 import { IWallet } from '@meshsdk/core';
 
 // Mock the txUtils module - fix the path
