+++ src/components/transaction-inspector/TransactionInspector.tsx
@@ -7,19 +7,19 @@ import { TextField, Box, Typography, Container, Paper, FormControlLabel, Checkbo
 import ContentCopyIcon from '@mui/icons-material/ContentCopy';
 import * as CSL from "@emurgo/cardano-serialization-lib-browser";
 import ReactJsonPretty from "react-json-pretty";
-import * as voteTxValidationUtils from "../utils/txValidationUtils";
-import { TransactionChecks } from "./txValidationChecks";
-import { VoteTransactionChecks } from "./voteValidationChecks";
-import { decodeHexToTx, convertGAToBech, getCardanoScanURL } from "../utils/txUtils";
-import { VotingDetails } from "./votingDetails";
-import { HierarchyDetails } from "./hierarchyDetails";
-import DownloadButton from "./molecules/downloadFiles";
-import FileUploader from "./molecules/fileUploader";
-import {TxValidationState,VoteTransactionDetails,VoteValidationState} from "./types/types";
-import {defaultTxValidationState,defaultVoteTransactionDetails,defaultVoteValidationState} from "./types/defaultStates";
-import SignTransactionButton from "./signTransactionButton";
-import TransactionDetailsActions from "./molecules/transactionDetailsActions";
-import txWitnessTemplate from "../../templates/cardano-file-templates/txWitnessTemplate.json";
+import * as voteTxValidationUtils from "../../utils/validation";
+import { ValidationChecks as TransactionChecks } from "./ValidationChecks";
+import { VoteTransactionChecks } from "./VoteValidationChecks";
+import { decodeHexToTx, convertGAToBech, getCardanoScanURL } from "../../utils/cardano";
+import { VotingDetails } from "./VotingDetails";
+import { HierarchyDetails } from "./HierarchyDetails";
+import DownloadButton from "../shared/downloadFiles";
+import FileUploader from "../shared/fileUploader";
+import {TxValidationState,VoteTransactionDetails,VoteValidationState} from "../../types/types";
+import {defaultTxValidationState,defaultVoteTransactionDetails,defaultVoteValidationState} from "../../types/defaultStates";
+import { SignTransaction as SignTransactionButton } from "./SignTransaction";
+import TransactionDetailsActions from "../shared/transactionDetailsActions";
+import txWitnessTemplate from "../../lib/templates/cardano-file-templates/txWitnessTemplate.json";
 
 export const TransactionButton = ({ 
   pendingTransactionHex,
