+++ src/components/wallet/WalletDetails.tsx
