+++ src/lib/templates/cardano-file-templates/cip136-template.json
