+++ src/lib/templates/cardano-file-templates/txWitnessTemplate.json
